i reorganized the entire db interaction code so i understood it
