+++ .idea/dictionaries/project.xml
@@ -3,6 +3,7 @@
     <words>
       <w>Joysound</w>
       <w>joysound</w>
+      <w>remocon</w>
     </words>
   </dictionary>
 </component>
