+++ src/server/kf2.toml
@@ -2,8 +2,7 @@
 path = "kf2.db"
 
 [server]
-host = "127.0.0.1"
-port = 3000
+listen_addr = "127.0.0.1:3000"
 
 [projector]
 root = "src/frontend/packages/projector/dist"
