+++ src/server/src/main.rs
@@ -1,7 +1,8 @@
-use server::config::load_config;
+use server::cli::parse_cli_args;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
-    let config = load_config().map_err(|e| anyhow::anyhow!("Failed to load configuration: {e}"))?;
+    let config =
+        parse_cli_args().map_err(|e| anyhow::anyhow!("Failed to load configuration: {e}"))?;
     server::run(config).await
 }
