+++ src/server/src/models.rs
@@ -1,0 +1,3 @@
+mod session;
+
+pub use session::Session;
